Fix height in Chrome
The heights of the images get computed when jQuery.ready fires.
Depending on network conditions, parser speed, parallelization, rendering timing etc., this can be before or after the images (or more precisely, the images' dimensions) have been loaded. Firefox seems to wait a little bit, whereas Chrome is more aggressive and renders before the images have been loaded.
On Chrome / If the height of the first image is not known by the time the carousel gets initialized, itemDefaultHeight will return 0, and the carousel will be invisible.